requirements-v3.8.0.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
